removed redundant tests from after spec, unified after decorator tests to a single file, added a test to make sure warnings do not fire outside of devMode
